+++ src/utils/EditableBMLabelProxy.cpp
@@ -1,6 +1,6 @@
 #include "EditableBMLabelProxy.hpp"
 
-EditableBMLabelProxy* EditableBMLabelProxy::create() {
+EditableBMLabelProxy* EditableBMLabelProxy::create(const char* font) {
     auto ret = new EditableBMLabelProxy();
     if (ret && ret->init()) {
         ret->autorelease();
@@ -12,34 +12,63 @@ EditableBMLabelProxy* EditableBMLabelProxy::create() {
 
 EditableBMLabelProxy* EditableBMLabelProxy::replace(
     CCLabelBMFont* existing,
+    CCNode* inputParent,
     float width, std::string const& placeholder,
-    std::function<void(std::string const&)> setValue
+    std::function<void(std::string const&)> onSetValue,
+    std::function<void(std::string const&)> onUpdate
 ) {
-    auto proxy = EditableBMLabelProxy::create();
+    auto proxy = EditableBMLabelProxy::create(existing->getFntFile());
 
-    proxy->m_setValue = setValue;
+    proxy->m_onSetValue = onSetValue;
+    proxy->m_onUpdate = onUpdate;
+    proxy->m_inputParent = inputParent;
     proxy->m_input = InputNode::create(width, placeholder.c_str());
     proxy->m_input->ignoreAnchorPointForPosition(false);
     proxy->m_input->getInput()->setDelegate(proxy);
     // 2021 Fod was sadistic for not doing this
     proxy->m_input->setContentSize(proxy->m_input->getBG()->getScaledContentSize());
-    existing->getParent()->addChild(proxy->m_input);
+    proxy->m_inputParent->addChild(proxy->m_input);
+
+    auto parent = existing->getParent();
+    parent->insertBefore(proxy, existing);
 
     proxy->setString(existing->getString());
     proxy->setScale(existing->getScale());
     proxy->setZOrder(existing->getZOrder());
+    proxy->setOrderOfArrival(existing->getOrderOfArrival());
     proxy->setPosition(existing->getPosition());
     proxy->setAnchorPoint(existing->getAnchorPoint());
     proxy->setID(existing->getID());
     proxy->setTag(existing->getTag());
+    proxy->setVisible(existing->isVisible());
 
-    existing->getParent()->addChild(proxy);
     existing->removeFromParent();
     
     return proxy;
 }
 
-void EditableBMLabelProxy::updateLabel() {}
+void EditableBMLabelProxy::setPosition(CCPoint const& pos) {
+    CCLabelBMFont::setPosition(pos);
+    if (m_input) m_input->setPosition(
+        m_inputParent->convertToNodeSpace(this->getParent()->convertToWorldSpace(m_obPosition))
+    );
+}
+
+void EditableBMLabelProxy::setScale(float scale) {
+    CCLabelBMFont::setScale(scale);
+    if (m_input) m_input->setScale(scale * 1.15f);
+}
+
+void EditableBMLabelProxy::setColor(ccColor3B const& color) {
+    CCLabelBMFont::setColor(color);
+    if (m_input) m_input->getInput()->setLabelNormalColor(color);
+}
+
+void EditableBMLabelProxy::updateLabel() {
+    if (m_input) {
+        this->setContentSize(m_input->getScaledContentSize());
+    }
+}
 
 void EditableBMLabelProxy::setString(const char* str) {
     this->setString(str, true);
@@ -50,14 +79,17 @@ void EditableBMLabelProxy::setString(const char* str, bool needUpdateLabel) {
     if (m_input && !m_ignoreLabelUpdate) {
         m_ignoreLabelUpdate = true;
         m_input->setString(str);
+        if (m_onUpdate) {
+            m_onUpdate(str);
+        }
         m_ignoreLabelUpdate = false;
     }
 }
 
 void EditableBMLabelProxy::textChanged(CCTextInputNode*) {
-    if (m_setValue && !m_ignoreLabelUpdate) {
+    if (m_onSetValue && !m_ignoreLabelUpdate) {
         m_ignoreLabelUpdate = true;
-        m_setValue(m_input->getString());
+        m_onSetValue(m_input->getString());
         m_ignoreLabelUpdate = false;
     }
 }