+++ src/utils/EditableBMLabelProxy.hpp
@@ -8,16 +8,20 @@ using namespace geode::prelude;
 class EditableBMLabelProxy : public CCLabelBMFont, TextInputDelegate {
 protected:
     Ref<InputNode> m_input = nullptr;
-    std::function<void(std::string const&)> m_setValue = nullptr;
+    Ref<CCNode> m_inputParent = nullptr;
+    std::function<void(std::string const&)> m_onSetValue = nullptr;
+    std::function<void(std::string const&)> m_onUpdate = nullptr;
     bool m_ignoreLabelUpdate = false;
     
-    static EditableBMLabelProxy* create();
+    static EditableBMLabelProxy* create(const char* font);
 
 public:
     static EditableBMLabelProxy* replace(
         CCLabelBMFont* existing,
+        CCNode* inputParent,
         float width, std::string const& placeholder,
-        std::function<void(std::string const&)> setValue
+        std::function<void(std::string const&)> onSetValue = nullptr,
+        std::function<void(std::string const&)> onUpdate = nullptr
     );
 
 #define PROXY_FWD_1(fun, param_ty) \
@@ -28,17 +32,20 @@ public:
 
     // PROXY_FWD_1(setContentSize, CCSize const&);
     PROXY_FWD_1(setAnchorPoint, CCPoint const&);
-    PROXY_FWD_1(setPosition, CCPoint const&);
     PROXY_FWD_1(setRotation, float);
     PROXY_FWD_1(setRotationX, float);
     PROXY_FWD_1(setRotationY, float);
-    PROXY_FWD_1(setScale, float);
     PROXY_FWD_1(setScaleX, float);
     PROXY_FWD_1(setScaleY, float);
-    PROXY_FWD_1(setVisible, bool);
     PROXY_FWD_1(setSkewX, float);
     PROXY_FWD_1(setSkewY, float);
+    PROXY_FWD_1(setVisible, bool);
     PROXY_FWD_1(setZOrder, int);
+    PROXY_FWD_1(setOrderOfArrival, unsigned int);
+
+    void setPosition(CCPoint const& pos) override;
+    void setScale(float scale) override;
+    void setColor(ccColor3B const& color) override;
 
     void updateLabel() override;
     void setString(const char* str) override;